added project files
This project is a sentiment analysis tool designed for Twitter. It employs a Convolutional Neural Network (CNN) model trained on Twitter data to predict the sentiment of user-inputted tweets. Using Streamlit as the interface, users can enter a tweet which undergoes tokenization and padding before being fed into the pre-trained CNN model. The model then predicts the sentiment as positive 😊, neutral 🙂, or negative 😖, based on the computed sentiment score. Additionally, the application displays a slider indicating the confidence level of the prediction. The interface is enhanced with a Lottie animation, providing an engaging visual component. The model is loaded alongside the tokenizer and sequence length parameters required for preprocessing the input text, offering real-time sentiment analysis for Twitter content.